+++ packages/bridge-status-controller/src/bridge-status-controller.test.ts
@@ -295,16 +295,8 @@ const getMockStartPollingForBridgeTxStatusArgs = ({
 } = {}): StartPollingForBridgeTxStatusArgsSerialized => ({
   bridgeTxMeta: {
     id: txMetaId,
+    hash: srcTxHash === 'undefined' ? undefined : srcTxHash,
   } as TransactionMeta,
-  statusRequest: {
-    bridgeId: 'lifi',
-    srcTxHash,
-    bridge: 'across',
-    srcChainId,
-    destChainId,
-    quote: getMockQuote({ srcChainId, destChainId }),
-    refuel: false,
-  },
   quoteResponse: {
     quote: getMockQuote({ srcChainId, destChainId }),
     trade: {
@@ -1077,8 +1069,9 @@ describe('BridgeStatusController', () => {
       });
 
       // Start polling with args that have no srcTxHash
-      const startPollingArgs = getMockStartPollingForBridgeTxStatusArgs();
-      startPollingArgs.statusRequest.srcTxHash = undefined;
+      const startPollingArgs = getMockStartPollingForBridgeTxStatusArgs({
+        srcTxHash: 'undefined',
+      });
       bridgeStatusController.startPollingForBridgeTxStatus(startPollingArgs);
 
       // Advance timer to trigger polling
@@ -1093,7 +1086,7 @@ describe('BridgeStatusController', () => {
       expect(
         bridgeStatusController.state.txHistory.bridgeTxMetaId1.status.srcChain
           .txHash,
-      ).toBeUndefined();
+      ).toBeFalsy();
 
       // Cleanup
       jest.restoreAllMocks();
@@ -1229,8 +1222,9 @@ describe('BridgeStatusController', () => {
       });
 
       // Start polling with no srcTxHash
-      const startPollingArgs = getMockStartPollingForBridgeTxStatusArgs();
-      startPollingArgs.statusRequest.srcTxHash = undefined;
+      const startPollingArgs = getMockStartPollingForBridgeTxStatusArgs({
+        srcTxHash: 'undefined',
+      });
       bridgeStatusController.startPollingForBridgeTxStatus(startPollingArgs);
 
       // Verify initial state has no srcTxHash
@@ -1843,7 +1837,8 @@ describe('BridgeStatusController', () => {
     beforeEach(() => {
       jest.clearAllMocks();
       jest.clearAllTimers();
-      jest.spyOn(Date, 'now').mockReturnValue(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567891);
       mockMessengerCall = jest.fn();
       mockMessengerCall.mockImplementationOnce(jest.fn()); // stopPollingForQuotes
     });
@@ -1852,6 +1847,10 @@ describe('BridgeStatusController', () => {
       mockMessengerCall.mockReturnValueOnce(mockSolanaAccount);
       mockMessengerCall.mockImplementationOnce(jest.fn()); // track event
       mockMessengerCall.mockResolvedValueOnce('signature');
+      mockMessengerCall.mockReturnValueOnce(mockSolanaAccount);
+      mockMessengerCall.mockReturnValueOnce({
+        transactions: [],
+      });
 
       const { controller, startPollingForBridgeTxStatusSpy } =
         getController(mockMessengerCall);
@@ -2055,7 +2054,8 @@ describe('BridgeStatusController', () => {
       jest.clearAllMocks();
       jest.clearAllTimers();
       mockMessengerCall = jest.fn();
-      jest.spyOn(Date, 'now').mockReturnValue(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567891);
       mockMessengerCall.mockImplementationOnce(jest.fn()); // stopPollingForQuotes
     });
 
@@ -2065,6 +2065,7 @@ describe('BridgeStatusController', () => {
       mockMessengerCall.mockResolvedValueOnce({
         signature: 'signature',
       });
+      mockMessengerCall.mockReturnValueOnce(mockSolanaAccount);
       mockMessengerCall.mockReturnValueOnce({
         transactions: [],
       });
@@ -2279,7 +2280,9 @@ describe('BridgeStatusController', () => {
     beforeEach(() => {
       jest.clearAllMocks();
       jest.clearAllTimers();
-      jest.spyOn(Date, 'now').mockReturnValue(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567891);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567892);
       mockMessengerCall = jest.fn();
       mockMessengerCall.mockImplementationOnce(jest.fn()); // stopPollingForQuotes
     });
@@ -2465,8 +2468,13 @@ describe('BridgeStatusController', () => {
       jest.clearAllMocks();
       jest.clearAllTimers();
       mockMessengerCall = jest.fn();
-      jest.spyOn(Date, 'now').mockReturnValue(1234567890);
-      jest.spyOn(Math, 'random').mockReturnValue(0.456);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567891);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567892);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567893);
+      jest.spyOn(Math, 'random').mockReturnValueOnce(0.456);
+      jest.spyOn(Math, 'random').mockReturnValueOnce(0.457);
+      jest.spyOn(Math, 'random').mockReturnValueOnce(0.458);
       mockMessengerCall.mockImplementationOnce(jest.fn()); // stopPollingForQuotes
     });
 
@@ -3175,7 +3183,7 @@ describe('BridgeStatusController', () => {
         ).toBeUndefined();
         expect(
           controller.state.txHistory[mockActionId].status.srcChain.txHash,
-        ).toBe(''); // Empty since tx was never submitted
+        ).toBeUndefined(); // Empty since tx was never submitted
 
         expect(startPollingForBridgeTxStatusSpy).toHaveBeenCalledTimes(0);
       });
@@ -3308,8 +3316,11 @@ describe('BridgeStatusController', () => {
     beforeEach(() => {
       jest.clearAllMocks();
       mockMessengerCall = jest.fn();
-      jest.spyOn(Date, 'now').mockReturnValue(1234567890);
-      jest.spyOn(Math, 'random').mockReturnValue(0.456);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567890);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567891);
+      jest.spyOn(Date, 'now').mockReturnValueOnce(1234567892);
+      jest.spyOn(Math, 'random').mockReturnValueOnce(0.456);
+      jest.spyOn(Math, 'random').mockReturnValueOnce(0.457);
       mockMessengerCall.mockImplementationOnce(jest.fn()); // stopPollingForQuotes
     });
 
@@ -4591,8 +4602,6 @@ describe('BridgeStatusController', () => {
         const unknownTxMetaId = 'unknown-tx-meta-id';
 
         const messengerCallSpy = jest.spyOn(mockBridgeStatusMessenger, 'call');
-
-        // Publish failure with an unknown txMeta.id but with matching actionId
         mockMessenger.publish('TransactionController:transactionFailed', {
           error: 'tx-error',
           transactionMeta: {
@@ -4875,6 +4884,10 @@ describe('BridgeStatusController', () => {
 
       it('should not append auth token to status request when getBearerToken throws an error', async () => {
         const messengerCallSpy = jest.spyOn(mockBridgeStatusMessenger, 'call');
+        consoleFnSpy = jest
+          .spyOn(console, 'error')
+          .mockImplementationOnce(jest.fn());
+        consoleFnSpy.mockImplementationOnce(jest.fn());
 
         messengerCallSpy.mockImplementation(() => {
           throw new Error(
@@ -4915,7 +4928,18 @@ describe('BridgeStatusController', () => {
             headers: { 'X-Client-Id': BridgeClientId.EXTENSION },
           },
         );
-        expect(consoleFnSpy).not.toHaveBeenCalled();
+        expect(consoleFnSpy.mock.calls).toMatchInlineSnapshot(`
+          [
+            [
+              "Error getting JWT token for bridge-api request",
+              [Error: AuthenticationController:getBearerToken not implemented],
+            ],
+            [
+              "Error getting JWT token for bridge-api request",
+              [Error: AuthenticationController:getBearerToken not implemented],
+            ],
+          ]
+        `);
       });
     });
   });