+++ packages/bridge-status-controller/src/utils/bridge-status.ts
@@ -1,5 +1,5 @@
 import { getClientHeaders } from '@metamask/bridge-controller';
-import type { Quote } from '@metamask/bridge-controller';
+import type { Quote, QuoteResponse } from '@metamask/bridge-controller';
 import { StructError } from '@metamask/superstruct';
 
 import { validateBridgeStatusResponse } from './validators';
@@ -10,6 +10,7 @@ import type {
   StatusRequestDto,
   FetchFunction,
   BridgeHistoryItem,
+  StatusRequest,
 } from '../types';
 
 export const getBridgeStatusUrl = (bridgeApiBaseUrl: string): string =>
@@ -109,4 +110,22 @@ export const shouldSkipFetchDueToFetchFailures = (
     }
   }
   return false;
+};
+
+/**
+ * @deprecated Use getStatusRequestWithSrcTxHash instead
+ * @param quoteResponse - The quote response to get the status request parameters from
+ * @returns The status request parameters
+ */
+export const getStatusRequestParams = (
+  quoteResponse: QuoteResponse,
+): StatusRequest => {
+  return {
+    bridgeId: quoteResponse.quote.bridgeId,
+    bridge: quoteResponse.quote.bridges[0],
+    srcChainId: quoteResponse.quote.srcChainId,
+    destChainId: quoteResponse.quote.destChainId,
+    quote: quoteResponse.quote,
+    refuel: Boolean(quoteResponse.quote.refuel),
+  };
 };